New: Added skinned mesh rendering, implemented as part of skinned.ppl pipeline New: Added new vertex type that contains skinning information (2 bones per vert) New: Added a new class AnimationTest that handles generating a test entity with a skinned cuboid-skeleton mesh to test the rendering logic Update: Pausing now affects physics simulation and animation updating Update: Added inverse bind pose matrices to the skeleton, required for skinned mesh rendering Update: Added extra rotation along Z axis in the test animation clip, to help test skinned mesh bone weights Cleanup: Model's upload descriptor no longer stores how many primitives there are (since it's a rendering, not a single-object-geometry concept) Bugfix: Model now correctly identifies whether it has indices when uploading

+++ Source/Graphics/Descriptor.cpp
@@ -17,7 +17,7 @@ void Descriptor::RecomputeSize()
 	const size_t size = myEntries.size();
 
 	myTotalSize = 0;
-	for (size_t i = 0; i < size; i++)
+	for (uint32_t i = 0; i < size; i++)
 	{
 		size_t basicAlignment = 0;
 		size_t elemSize = 0;
